+++ misc/misc.h
@@ -10,4 +10,4 @@ uint ipow(uint a, uint b);
 // parses int at start and stores result in result;
 // returns length of the string representation of the number.
 uint parse_uint(char* start, uint* result);
-uint parse_int(char* start, int* result);
+uint parse_int(char* start, long* result);
